ride-service/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
pika
python-dotenv
djangorestframework-simplejwt
gunicorn
django-cors-headers
